+++ Labs10/Labs10.cpp
@@ -36,81 +36,56 @@
     *
 */
 #include <iostream>
-#include "MatrixBase.h"
+#include "Matrix2D.h"
+#include "Matrix3D.h"
 
 using namespace std;
 
-ostream& operator<<(ostream& stream, const MatrixBase& iMatrix) {
-   for (int i = 0; i < iMatrix.m_size; i++) {
-      for (int j = 0; j < iMatrix.m_size; j++) {
-         stream << iMatrix.matrix[i][j] << " ";
-      }
-      stream << endl;
-   }
-   return stream;
-}
-
-void MatrixBase::operator*=(int iMult) {
-   for (int i = 0; i < m_size; i++) {
-      for (int j = 0; j < m_size; j++) {
-         matrix[i][j] *= iMult;
-      }
-   }
-}
-
-void MatrixBase::operator+=(MatrixBase &iAdd) {
-   for (int i = 0; i < m_size; i++) {
-      for (int j = 0; j < m_size; j++) {
-         matrix[i][j] += iAdd.element(i, j);
-      }
-   }
-}
-
 void create(MatrixBase matrix) {
-   int num = 1;
-   for (int i = 0; i < matrix.size(); i++) {
-      for (int j = 0; j < matrix.size(); j++) {
-         matrix.element(i, j) = num;
-         num++;
-      }
-   }
+    int num = 1;
+    for (int i = 0; i < matrix.size(); i++) {
+        for (int j = 0; j < matrix.size(); j++) {
+            matrix.element(i, j) = num;
+            num++;
+        }
+    }
 }
 
 int main()
 {
-   Matrix2D m2a, m2b;
-   Matrix3D m3a, m3b;
+    Matrix2D m2a, m2b;
+    Matrix3D m3a, m3b;
 
-   create(m2a);
-   create(m2b);
-   create(m3a);
-   create(m3b);
-   cout << "Матрицы после заполнения значениями" << endl;
-   cout << "Матрица m2a" << endl;
-   cout << m2a << endl;
-   cout << "Матрица m2b" << endl;
-   cout << m2b << endl;
-   cout << "Матрица m3a" << endl;
-   cout << m3a << endl;
-   cout << "Матрица m3b" << endl;
-   cout << m3b << endl;
+    create(m2a);
+    create(m2b);
+    create(m3a);
+    create(m3b);
+    cout << "Матрицы после заполнения значениями" << endl;
+    cout << "Матрица m2a" << endl;
+    cout << m2a << endl;
+    cout << "Матрица m2b" << endl;
+    cout << m2b << endl;
+    cout << "Матрица m3a" << endl;
+    cout << m3a << endl;
+    cout << "Матрица m3b" << endl;
+    cout << m3b << endl;
 
    // Умножение матриц на число
    m2a *= 2;
    m3a *= 5;
    // Сложение матриц
-   m2b += m2a;
-   m3b += m3b;
+    m2b += m2a;
+    m3b += m3b;
 
-   cout << "Матрицы после проведения операций сложения  и умножения" << endl;
-   cout << "Было произведено умножения на 2" << endl;
-   cout << m2a << endl;
-   cout << "Сложили матрицу m2b с матрицей m2a после умножения m2a на 2" << endl;
-   cout << m2b << endl;
-   cout << "Матрица m3a была умножена на 5" << endl;
-   cout << m3a << endl;
-   cout << "Результат сложения двух матриц m3b" << endl;
-   cout << m3b << endl;
+    cout << "Матрицы после проведения операций сложения  и умножения" << endl;
+    cout << "Было произведено умножения на 2" << endl;
+    cout << m2a << endl;
+    cout << "Сложили матрицу m2b с матрицей m2a после умножения m2a на 2" << endl;
+    cout << m2b << endl;
+    cout << "Матрица m3a была умножена на 5" << endl;
+    cout << m3a << endl;
+    cout << "Результат сложения двух матриц m3b" << endl;
+    cout << m3b << endl;
 
-   return 0;
+    return 0;
 }
